A lot of UNTESTED changes -- see description
- Implemented the edit action for OM
- the OM form page accepts a required prop (state) which will define how the page will be rendered (new request/edit request/modify the request in case of return)

editing algorithm --> when you click on edit/Modify, it sets the OM identity, loads the OM details, change the page content, and render the FORM page with the details prefilled.
save as draft wont be there in case of modify.
this has only been fully implemented for OM (needs test tho). For AC and DC, it still to trigger the loading of details in useEffect and handleOnClick on confirm and Submit, and need to create the component for view.

- the snackbar notification in all tables pages pops up depending on the state of the status (saved, deleted, submitted...)